+++ api/v1/base_types.go
@@ -17,7 +17,10 @@ limitations under the License.
 package v1
 
 import (
+	machineryapi "github.com/cloudnative-pg/machinery/pkg/api"
 	corev1 "k8s.io/api/core/v1"
+
+	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 )
 
 // PodStatus represent the possible status of pods
@@ -33,6 +36,21 @@ const (
 	// PodFailed means that a Pod will not be scheduled again (deleted or evicted)
 	PodFailed = "failed"
 )
+
+// LocalObjectReference contains enough information to let you locate a
+// local object with a known type inside the same namespace
+// +kubebuilder:object:generate:=false
+type LocalObjectReference = machineryapi.LocalObjectReference
+
+// SecretKeySelector contains enough information to let you locate
+// the key of a Secret
+// +kubebuilder:object:generate:=false
+type SecretKeySelector = machineryapi.SecretKeySelector
+
+// ConfigMapKeySelector contains enough information to let you locate
+// the key of a ConfigMap
+// +kubebuilder:object:generate:=false
+type ConfigMapKeySelector = machineryapi.ConfigMapKeySelector
 
 // SecretKeySelectorToCore transforms a SecretKeySelector structure to the
 // analogue one in the corev1 namespace
@@ -62,4 +80,25 @@ func ConfigMapKeySelectorToCore(selector *ConfigMapKeySelector) *corev1.ConfigMa
 		},
 		Key: selector.Key,
 	}
+}
+
+// ListStatusPods return a list of active Pods
+func ListStatusPods(podList []corev1.Pod) map[PodStatus][]string {
+	podsNames := make(map[PodStatus][]string)
+
+	for _, pod := range podList {
+		if !pod.DeletionTimestamp.IsZero() {
+			continue
+		}
+		switch {
+		case utils.IsPodReady(pod):
+			podsNames[PodHealthy] = append(podsNames[PodHealthy], pod.Name)
+		case utils.IsPodActive(pod):
+			podsNames[PodReplicating] = append(podsNames[PodReplicating], pod.Name)
+		default:
+			podsNames[PodFailed] = append(podsNames[PodFailed], pod.Name)
+		}
+	}
+
+	return podsNames
 }
