+++ api/v1/base_types_test.go
@@ -17,6 +17,9 @@ limitations under the License.
 package v1
 
 import (
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
@@ -54,5 +57,109 @@ var _ = Describe("Base type mappings for configmaps", func() {
 
 		Expect(selector.Name).To(Equal("thisName"))
 		Expect(selector.Key).To(Equal("thisKey"))
+	})
+})
+
+var _ = Describe("Pod conditions test suite", func() {
+	Describe("Properly builds ListStatusPods", func() {
+		healthyPod := corev1.Pod{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "healthyPod",
+			},
+			Status: corev1.PodStatus{
+				Conditions: []corev1.PodCondition{
+					{
+						Type:   corev1.ContainersReady,
+						Status: corev1.ConditionTrue,
+					},
+				},
+			},
+		}
+		activePod := corev1.Pod{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "activePod",
+			},
+			Status: corev1.PodStatus{
+				Phase: corev1.PodRunning,
+				Conditions: []corev1.PodCondition{
+					{
+						Type:   corev1.ContainersReady,
+						Status: corev1.ConditionFalse,
+					},
+				},
+			},
+		}
+		failedPod := corev1.Pod{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "failedPod",
+			},
+			Status: corev1.PodStatus{
+				Phase: corev1.PodFailed,
+				Conditions: []corev1.PodCondition{
+					{
+						Type:   corev1.ContainersReady,
+						Status: corev1.ConditionFalse,
+					},
+				},
+			},
+		}
+
+		now := metav1.Now()
+		terminatingPod := corev1.Pod{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:              "terminatingPod",
+				DeletionTimestamp: &now,
+			},
+			Status: corev1.PodStatus{
+				Phase: corev1.PodRunning,
+				Conditions: []corev1.PodCondition{
+					{
+						Type:   corev1.ContainersReady,
+						Status: corev1.ConditionTrue,
+					},
+				},
+			},
+		}
+
+		It("Detects healthy pods", func() {
+			podList := []corev1.Pod{healthyPod, healthyPod}
+			expectedStatus := map[PodStatus][]string{
+				PodHealthy: {"healthyPod", "healthyPod"},
+			}
+			podStatus := ListStatusPods(podList)
+			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
+		})
+
+		It("Detects active pods", func() {
+			podList := []corev1.Pod{healthyPod, activePod}
+			expectedStatus := map[PodStatus][]string{
+				PodHealthy:     {"healthyPod"},
+				PodReplicating: {"activePod"},
+			}
+			podStatus := ListStatusPods(podList)
+			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
+		})
+
+		It("Detects failed pods", func() {
+			podList := []corev1.Pod{healthyPod, activePod, failedPod}
+			expectedStatus := map[PodStatus][]string{
+				PodHealthy:     {"healthyPod"},
+				PodReplicating: {"activePod"},
+				PodFailed:      {"failedPod"},
+			}
+			podStatus := ListStatusPods(podList)
+			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
+		})
+
+		It("Excludes terminating pods", func() {
+			podList := []corev1.Pod{healthyPod, activePod, failedPod, terminatingPod}
+			expectedStatus := map[PodStatus][]string{
+				PodHealthy:     {"healthyPod"},
+				PodReplicating: {"activePod"},
+				PodFailed:      {"failedPod"},
+			}
+			podStatus := ListStatusPods(podList)
+			Expect(podStatus).To(BeEquivalentTo(expectedStatus))
+		})
 	})
 })
